refactor: introduce `proto_reader.rs`
